Improved Label filtering

+++ Actions/filter.py
@@ -1,12 +1,24 @@
 #!/usr/bin/env python
+# -*- coding: utf-8 -*-
 
 import logging
 from lxml import etree
 
 class LabelFilter(object):
     
+    def process_most_popular_labels(self, count):
+        
+        # Sanity checks
+        assert count is not None
+        count = int(count)
+        assert count > 0
+        
+        self.using_most_popular_labels = True 
+        self.number_of_popular_labels = count
+    
     def __init__(self, xml):
        self.labels = set([])
+       self.using_most_popular_labels = False
        
        # Pull out all of the labels we're keeping
        for x_node in xml.getchildren():
@@ -14,6 +26,8 @@ class LabelFilter(object):
                continue
            elif x_node.tag == "RetainLabel":
                self.labels.add(int(x_node.get("id")))
+           elif x_node.tag == "RetainMostPopularLabels":
+               self.process_most_popular_labels(x_node.get("count"))
            else:
                raise ValueError((x_node.tag, "Unsupported here"))
        
@@ -35,9 +49,6 @@ class LabelFilter(object):
            self.delete = False
        else:
            self.delete = True
-           
-       if len(self.labels) == 0:
-           raise ValueError("Need some labels to retain")
     
     def get_table(self):
         # Create the string 
@@ -48,6 +59,8 @@ class LabelFilter(object):
         return table 
     
     def drop_identifiers(self, cursor):
+        if len(self.labels) == 0:
+            raise ValueError("Need some labels to retain")
         table = self.get_table()
         not_portion = ','.join([str(i) for i in self.labels])
         sql = r"""DELETE FROM input WHERE identifier IN (
@@ -59,10 +72,31 @@ class LabelFilter(object):
         logging.info("Deleting non-matching documents...")
         cursor.execute(sql)
     
+    def determine_most_popular_labels(self, cursor):
+        
+        assert self.using_most_popular_labels
+        
+        # Build query 
+        sql = "SELECT COUNT(*) AS c, label FROM %s GROUP BY label ORDER BY c DESC LIMIT 0, %d"
+        table = self.get_table()
+        
+        # Execute query 
+        cursor.execute(sql % (table, self.number_of_popular_labels))
+        
+        for popularity, label in cursor.fetchall():
+            logging.debug("Discovered new label '%d' with popularity '%d'", label, popularity)
+            yield label 
+    
     def execute(self, path, conn):
         
         # Grab a cursor
         c = conn.cursor()
+        
+        # Check if we need to generate any more labels
+        if self.using_most_popular_labels:
+            logging.info("Determining the %d most popular labels...", self.number_of_popular_labels)
+            for label in self.determine_most_popular_labels(c):
+                self.labels.add(label)
     
         # Delete the documents
         self.drop_identifiers(c)
